Fix xrGameLA, part 13

+++ src/xrGameLA/debug_renderer.cpp
@@ -12,7 +12,7 @@
 
 #include "debug_renderer.h"
 
-void CDebugRenderer::add_lines		(Fvector const *vertices, u32 const &vertex_count, u16 const *pairs, u32 const &pair_count, u32 const &color)
+void CDebugRenderer::add_lines		(Fvector const *vertices, u32 const &vertex_count, u32 const *pairs, u32 const &pair_count, u32 const &color)
 {
 	DRender->add_lines				(vertices, vertex_count, pairs, pair_count, color);
 }
@@ -29,7 +29,7 @@ void CDebugRenderer::draw_obb		(const Fmatrix &matrix, const u32 &color)
 	matrix.transform_tiny			(aabb[6],Fvector().set( +1, +1, +1)); // 6
 	matrix.transform_tiny			(aabb[7],Fvector().set( +1, -1, +1)); // 7
 
-	u16								aabb_id[12*2] = {
+	u32								aabb_id[12*2] = {
 		0,1,  1,2,  2,3,  3,0,  4,5,  5,6,  6,7,  7,4,  1,5,  2,6,  3,7,  0,4
 	};
 
@@ -89,7 +89,7 @@ void CDebugRenderer::draw_ellipse	(const Fmatrix &matrix, const u32 &color)
 			0.2706f,0.2706f,-0.9239f,  0.1464f,0.3536f,-0.9239f,  0.0000f,0.0000f,-1.0000f
 	};
 
-	u16 pairs[]						= {
+	u32 pairs[]						= {
 		0,1, 0,2, 0,3, 0,4, 0,5, 0,6, 0,7, 0,8, 0,9, 0,10, 0,11, 0,12, 0,13, 0,14, 0,15,
 		0,16, 1,2, 1,17, 1,18, 2,3, 2,18, 2,19, 3,4, 3,19, 3,20, 4,5, 4,20, 4,21, 5,6,
 		5,21, 5,22, 6,7, 6,22, 6,23, 7,8, 7,23, 7,24, 8,9, 8,24, 8,25, 9,10, 9,25, 9,26,
